More refactoring of helpers, added stub for visualiser
